+++ runtime/src/main/java/com/flipkart/flux/initializer/FluxInitializer.java
@@ -112,6 +112,8 @@ public class FluxInitializer {
                 new TaskModule(),
                 new FluxClientInterceptorModule()
         );
+        //scans package com.flipkart.flux for polyguice specific annotations like @Bindable, @Component etc.
+        fluxRuntimeContainer.scanPackage("com.flipkart.flux");
         fluxRuntimeContainer.registerConfigurationProvider(configModule.getConfigProvider());
         fluxRuntimeContainer.prepare();
     }