+++ CHANGELOG.md
@@ -7,3 +7,11 @@
   - Http API interface for triggering workflows, inspection and monitoring
   - Real-time monitoring of Task execution across state machine instances
   - Visualization for inspecting individual State machine executions
+  
+## 1.0.3-SNAPSHOT (Oct 24, 2016)
+- **New features:**
+  - Support for Deployment Units
+  - Global redriver
+  - Workflows versioning support
+  - Support for Retriable exceptions
+  - Http API for unsidelining sidelined/errored tasks